More frontend work
Implemented new backend version, added the ability to view the traffic cam img, error handling. The issue when the player or computer ran out of cards was fixed thx to the new backend version. The card dealing issue due to the queryParam in the goFishAPI.java had to be resolved again.

Co-Authored-By: Andreas Persson <151462464+infotillandreas@users.noreply.github.com>